requirementx.txt
flake8==5.0.4
isort==5.10.1
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.20.0
